+++ constants/triggers.js
@@ -112,6 +112,9 @@ export const MEGALODON_PET_EPIC_MESSAGE = `PET DROP! ${RESET}${DARK_PURPLE}Megal
 export const DEEP_SEA_ORB_MESSAGE = `RARE DROP! ${RESET}${DARK_PURPLE}Deep Sea Orb ${MAGIC_FIND_MESSAGE_PATTERN}`;
 export const RADIOACTIVE_VIAL_MESSAGE = `RARE DROP! ${RESET}${LIGHT_PURPLE}Radioactive Vial ${MAGIC_FIND_MESSAGE_PATTERN}`; // RARE DROP! &r&dRadioactive Vial &r&b(+&r&b236% &r&b✯ Magic Find&r&b)
 export const MAGMA_CORE_MESSAGE = `RARE DROP! ${RESET}${BLUE}Magma Core ${MAGIC_FIND_MESSAGE_PATTERN}`; // RARE DROP! &r&9Magma Core &r&b(+&r&b236% &r&b✯ Magic Find&r&b)
+export const TIKI_MASK_MESSAGE = `RARE DROP! ${RESET}${GOLD}Tiki Mask ${MAGIC_FIND_MESSAGE_PATTERN}`; // RARE DROP! &r&6Tiki Mask &r&b(+&r&b236% &r&b✯ Magic Find&r&b)
+export const TITANOBOA_SHED_MESSAGE = `RARE DROP! ${RESET}${GOLD}Titanoboa Shed ${MAGIC_FIND_MESSAGE_PATTERN}`; // RARE DROP! &r&6Titanoboa Shed &r&b(+&r&b236% &r&b✯ Magic Find&r&b)
+export const SCUTTLER_SHELL_MESSAGE = `RARE DROP! ${RESET}${GOLD}Scuttler Shell ${MAGIC_FIND_MESSAGE_PATTERN}`; // RARE DROP! &r&6Scuttler Shell &r&b(+&r&b236% &r&b✯ Magic Find&r&b)
 
 export const AQUAMARINE_DYE_MESSAGE = `${RESET}${LIGHT_PURPLE}${BOLD}WOW! ` + '${playerNameAndRank}' + ` ${RESET}${GOLD}found ${RESET}${AQUA}Aquamarine Dye`; // &r&d&lWOW! &r&b[MVP&r&c+&r&b] &bMoonTheSadFisher&r&r&f &r&6found &r&bAquamarine Dye &r&8#95&r&6!&r
 export const ICEBERG_DYE_MESSAGE = `${RESET}${LIGHT_PURPLE}${BOLD}WOW! ` + '${playerNameAndRank}' + ` ${RESET}${GOLD}found ${RESET}${DARK_AQUA}Iceberg Dye`; // &r&d&lWOW! &r&b[MVP&r&c+&r&b] &bMoonTheSadFisher&r&r&f &r&6found &r&3Iceberg Dye&r
@@ -762,6 +765,36 @@ export const RARE_DROP_TRIGGERS = [
         isAlertEnabledSettingKey: 'alertOnMagmaCoreDrop',
         rarityColorCode: RARE,
         shouldTrackDropNumber: true,
+    },
+    {
+        trigger: TIKI_MASK_MESSAGE,
+        itemId: 'TIKI_MASK',
+        itemName: drops.TIKI_MASK,
+        sound: sounds.MC_RARE_ACHIEVEMENT_SOURCE,
+        isMessageEnabledSettingKey: 'messageOnTikiMaskDrop',
+        isAlertEnabledSettingKey: 'alertOnTikiMaskDrop',
+        rarityColorCode: LEGENDARY,
+        shouldTrackDropNumber: false,
+    },
+    {
+        trigger: TITANOBOA_SHED_MESSAGE,
+        itemId: 'TITANOBOA_SHED',
+        itemName: drops.TITANOBOA_SHED,
+        sound: sounds.MC_RARE_ACHIEVEMENT_SOURCE,
+        isMessageEnabledSettingKey: 'messageOnTitanoboaShedDrop',
+        isAlertEnabledSettingKey: 'alertOnTitanoboaShedDrop',
+        rarityColorCode: LEGENDARY,
+        shouldTrackDropNumber: false,
+    },
+    {
+        trigger: SCUTTLER_SHELL_MESSAGE,
+        itemId: 'SCUTTLER_SHELL',
+        itemName: drops.SCUTTLER_SHELL,
+        sound: sounds.OH_MY_GOD_SOUND_SOURCE,
+        isMessageEnabledSettingKey: 'messageOnScuttlerShellDrop',
+        isAlertEnabledSettingKey: 'alertOnScuttlerShellDrop',
+        rarityColorCode: LEGENDARY,
+        shouldTrackDropNumber: false,
     },
 ];